# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Duel Standby API
  version: 0.4.0
  description: Auth, dashboard, users, teams, tournaments, treasury, profile, and audit endpoints for DSWeb.
servers:
  - url: https://example.com
components:
  securitySchemes:
    dsAuthCookie:
      type: apiKey
      in: cookie
      name: ds_auth
...
paths:
  /api/auth/register:
    post:
      summary: Register a new public user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
                - confirmPassword
                - phoneWhatsapp
                - city
                - sourceInfo
                - socialMedia
                - agreement
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                confirmPassword:
                  type: string
                phoneWhatsapp:
                  type: string
                city:
                  type: string
                duelLinksScreenshotUploadId:
                  type: string
                masterDuelScreenshotUploadId:
                  type: string
                sourceInfo:
                  type: string
                socialMedia:
                  type: array
                  items:
                    type: string
                agreement:
                  type: boolean
      responses:
        "201":
          description: Registration created
  /api/auth/login:
    post:
      summary: Legacy credentials login endpoint for compatibility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifier
                - password
              properties:
                identifier:
                  type: string
                  description: Username atau email
                password:
                  type: string
                rememberMe:
                  type: boolean
      responses:
        "200":
          description: Login success
        "401":
          description: Invalid credentials
  /api/auth/finalize:
    get:
      summary: Finalize Auth.js session into internal auth cookies and redirect
      responses:
        "302":
          description: Redirect to dashboard or requested internal path after issuing auth cookies
    post:
      summary: Finalize Auth.js session into internal auth cookies for client-side login
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  enum:
                    - credentials
                    - google
                redirect:
                  type: string
      responses:
        "200":
          description: Internal cookies issued and redirect target returned
        "401":
          description: Auth.js session missing or invalid
...
  /api/tournaments/{id}/register:
    post:
      summary: Register authenticated active user to a tournament
      security:
        - dsAuthCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Registration success
        "400":
          description: Registration rejected
        "401":
          description: Unauthorized
        "404":
          description: Not found